% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHAUDI.R
\name{fHAUDI}
\alias{fHAUDI}
\title{fHAUDI}
\usage{
fHAUDI(
  fbm_obj,
  fbm_anno_info,
  y,
  gamma_anno_s,
  family = "gaussian",
  ind_train = NULL,
  snps = NULL,
  ...
)
}
\arguments{
\item{fbm_obj}{object of FBM class}

\item{fbm_anno_info}{info data frame containing information FBM (chrom/pos/samples/etc.) merged with annotation data}

\item{y}{Vector of responses}

\item{gamma_anno_s}{Vector specifying multiplicative penalty for gamma and then annotation specific penalties}

\item{family}{Either "gaussian" or "binomial" (default is gaussian)}

\item{ind_train}{Vector of indices specifying the rows to use for training the model}

\item{snps}{Vector of SNPs to include in model}

\item{...}{additional arguments to pass to big_spLinReg / big_spLogReg}
}
\value{
An object of class \code{big_sp_list} from the \code{bigstatsr} package
}
\description{
Fit fHAUDI model
}
\author{
Brian Chen
}
